add Path object support
File path are now represented by Path objects instead of strings. In combination with transparent path this allows for seamless switching between google storage and local files.
With the optional datafolder parameter, the bag file can even be a file object or buffer object. This allows for integration with more foreign storage solutions